git: add wrapper script to get logs from git-tfs
Managing multiple repositories/projects in one git repo allows for
commits that span repos, but must be done as a subtree. This means that
the standard git-log on a file will only show the subtree merge
commits, rather than the history of each file. This wrapper munges
the arguments to point at the correct git-tfs remote, and file location.